[Refactor] Gacha System 코드 수정
- System 코드 리팩토링
