+++ Assets/Programing/YJE/Test_GachaScene_YJE.unity
@@ -833,7 +833,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1ec2f93d17defbe42b7bf1bbaa8172f0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _uID: Y29oJ7Tu2RQr0SZlbgYzZcDz5Xb2
+  _uID: CHSmbrwghYNzZb7AIkdLRtvpHaW2
   _resetFollowTime: 0
   _originFollowTime: 0
 --- !u!114 &327574269
